Mainpage files restructured
