+++ package.json
@@ -7,6 +7,7 @@
   "dependencies": {
     "bootstrap": "^4.3.1",
     "gatsby": "^2.12.0",
+    "gatsby-cli": "^2.7.47",
     "gatsby-image": "^2.2.3",
     "gatsby-plugin-manifest": "^2.2.0",
     "gatsby-plugin-offline": "^2.2.0",
