+++ src/pages/404.js
@@ -1,7 +1,7 @@
 import React from "react"
 
 import Layout from "../components/layout"
-import SEO from "../components/seo"
+import SEO from "../components/common/seo"
 
 const NotFoundPage = () => (
   <Layout>
