+++ src/pages/competicao-programacao.js
@@ -2,7 +2,7 @@ import React from "react"
 import { Link } from "gatsby"
 
 import Layout from "../components/layout"
-import SEO from "../components/seo"
+import SEO from "../components/common/seo"
 import Competition from "../components/competition"
 
 const CompetitionPage = () => (
