+++ src/pages/index.js
@@ -1,15 +1,16 @@
 import React from "react"
 import "bootstrap/dist/css/bootstrap.min.css"
 
-import SEO from "../components/seo"
-import Navbar from "../components/navbar"
+import SEO from "../components/common/seo"
+import Navbar from "../components/common/navbar"
 
-import Tickets from "../components/tickets"
-import Speakers from "../components/speakers"
-import Banner from "../components/mainpage/banner"
-import Contacts from "../components/contacts/wrapper"
+import About from "../components/mainpage/about/about"
+import Tickets from "../components/mainpage/tickets"
+import Speakers from "../components/mainpage/speakers"
+import Banner from "../components/mainpage/banner/banner"
+import Contacts from "../components/mainpage/contacts/wrapper"
 
-import MainpageStyles from "../styles/mainpage.module.css"
+import MainpageStyles from "../styles/mainpage/mainpage.module.css"
 
 const IndexPage = () => (
   <div>
@@ -21,12 +22,10 @@ const IndexPage = () => (
     <div className={MainpageStyles.sticky}>
       <Navbar logo />
     </div>
-    <Tickets />
+    <About />
     <Speakers />
-    <Contacts />
-    <Contacts />
-    <Contacts />
-    <Contacts />
+    <Tickets />
+    {/* <Sponsors /> */}
     <Contacts />
   </div>
 )
