+++ src/pages/program.js
@@ -2,7 +2,7 @@ import React from "react"
 import { graphql } from "gatsby"
 
 import Layout from "../components/layout"
-import SEO from "../components/seo"
+import SEO from "../components/common/seo"
 import DailySchedule from "../components/dailySchedule"
 import Event from "../components/event"
