+++ src/pages/speakers.js
@@ -3,7 +3,7 @@ import { graphql, Link } from "gatsby"
 import Img from "gatsby-image"
 
 import Layout from "../components/layout"
-import SEO from "../components/seo"
+import SEO from "../components/common/seo"
 import Occupations from "../components/occupations"
 
 const styles = {
