+++ src/styles/mainpage/speakers.module.css
@@ -1,3 +1,5 @@
+@import "../common/palette.css";
+
 .speakersSection {
   margin-top: 5rem;
   margin-bottom: 8rem;
