+++ src/styles/mainpage/tickets.module.css
@@ -1,4 +1,5 @@
 @import url("https://fonts.googleapis.com/css?family=Roboto:100,300,400,500&display=swap");
+@import "../common/palette.css";
 
 .wrapper {
   border: 4px var(--white) solid;
@@ -44,7 +45,8 @@
 
 .section {
   margin: 2rem 0;
-  background: url("../images/showcase_events_filter.jpg") center center / cover;
+  background: url("../../images/showcase_events_filter.jpg") center center /
+    cover;
   text-align: center;
   color: var(--white);
   padding: 10% 20%;
